feat: enhance IPC frame contract and compatibility tests
