+++ gems/vajra/spec/cpp/CMakeLists.txt
@@ -10,6 +10,9 @@ find_package(Threads REQUIRED)
 add_executable(vajra_server_test
   test_support.cpp
   ipc_contract_test.cpp
+  ipc_frame_contract_test.cpp
+  ipc_frame_header_test.cpp
+  ipc_protocol_compatibility_test.cpp
   server_lifecycle_test.cpp
   request_head_test.cpp
   response_test.cpp
