+++ gems/vajra/spec/cpp/test_suites.hpp
@@ -9,6 +9,9 @@
 namespace VajraSpecCpp
 {
   void run_ipc_contract_tests();
+  void run_ipc_frame_contract_tests();
+  void run_ipc_frame_header_tests();
+  void run_ipc_protocol_compatibility_tests();
   void run_server_lifecycle_tests();
   void run_request_head_tests();
   void run_response_tests();
